requirements.txt
fastapi==0.115.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
openai==1.51.0
pandas==2.2.3
numpy==1.26.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
streamlit==1.39.0
plotly==5.24.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pgvector==0.3.5
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
uvicorn==0.30.6
